requirements.txt
torch>=2.1.0
torchvision>=0.16.0
timm>=0.9.12
diffusers>=0.25.0
transformers>=4.36.0
accelerate>=0.25.0
einops>=0.7.0
numpy>=1.24.0
Pillow>=10.0.0
tqdm>=4.66.0
scipy>=1.11.0
pytorch-fid>=0.3.0
clean-fid>=0.1.35
wandb>=0.16.0
omegaconf>=2.3.0
matplotlib>=3.8.0
